+++ R/S3-methods.R
@@ -21,6 +21,17 @@ model.frame.gls <- function(formula, ...) {
   }
 }
 
+
+#' @export
+model.frame.gee <- function(formula, ...) {
+  # get function call
+  tmp <- as.list(formula$call)
+  model.vars <- all.vars(tmp$formula)
+  all.data <- get(deparse(tmp$data))
+  all.data[, model.vars]
+}
+
+
 #' @export
 print.sjstats_r2 <- function(x, ...) {
   s3 <- NULL
@@ -282,4 +293,32 @@ print.sjstats_zcf <- function(x, ...) {
 #' @export
 print.sjstats_outliers <- function(x, ...) {
   print(x$result, ...)
+}
+
+#' @export
+print.sj_xtab_stat <- function(x, ...) {
+  # get length of method name, to align output
+  l <- nchar(x$method)
+
+  # is method shorter than p-value?
+  if (l < 7) l <- 7
+
+  # headline
+  cat("Measure of Association for Contingency table\n")
+
+  # used fisher?
+  if (x$fisher)
+    cat("                 (using Fisher's Exact Test)\n")
+  else
+      cat("\n")
+
+  # print test statistic
+  cat(sprintf("  %*s: %.4f\n", l, x$stat.name, x$statistic))
+  cat(sprintf("  %*s: %.4f\n", l, x$method, x$estimate))
+
+  # check if p <.001
+  if (x$p.value < 0.001)
+    cat(sprintf("  %*s: <0.001\n", l, "p-value", x$p.value))
+  else
+    cat(sprintf("  %*s: %.4f\n", l, "p-value", x$p.value))
 }